cloudflared_tunnel/requirements.txt
fastapi
uvicorn
pyjwt
psutil
pyyaml
requests
qrcode[pil]